홈 레이아웃만 작업 완료 (로직 작업 필요)
